Update README.md
Hello reader! This is a collab with @Popper412-Geofs and @redlarvaoioi . This is a custom project! Please do NOT copy and paste this into your OWN repository and say that it's yours. If you do it. Please let either me and the other person know or give us FULL credit in your desc. Thank you. Enjoy! Issues and Pull requests are indeed open!